WIP: rework using syncified code. remove `sections.rs`, just have them as methods
