+++ src/gui/mod.rs
@@ -1,16 +1,18 @@
 mod app;
-mod sections;
 
 use app::App;
 use std::sync::Arc;
-use tokio::runtime::Handle;
 
-use crate::logging::LogPaths;
+use crate::{logging::LogPaths, orchestrator::Orchestrator, runtime::RemoteSpawn};
 
-pub fn main(log_paths: Arc<LogPaths>) -> eframe::Result<()> {
+pub fn main(
+    runtime: impl RemoteSpawn,
+    orc: Arc<impl Orchestrator + Send + Sync + 'static>,
+    log_paths: Arc<LogPaths>,
+) -> eframe::Result<()> {
     eframe::run_native(
         "caligula-gui",
         Default::default(),
-        Box::new(|cc| Ok(Box::new(App::new(cc, log_paths)))),
+        Box::new(|cc| Ok(Box::new(App::new(cc, runtime, orc, log_paths)))),
     )
 }
